Support importing stargz snapshotter as a plugin
This commit introduces `github.com/containerd/stargz-snapshotter/service/plugin`
pkg which can be importd as a *builtin* stargz snapshotter plugin.
This also adds integration tests for builtin stargz snapshotter.
